Beginner_Python_Project_Complete/requirements.txt
click==7.1.2
Flask==1.1.2
Flask-Materialize==0.2.4
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
Werkzeug==1.0.1
